Keep `bandwidth` and `indicatedBitrate` in bits per second

+++ pillarbox-player/src/main/java/ch/srgssr/pillarbox/player/qos/models/QoSEvent.kt
@@ -7,8 +7,8 @@ package ch.srgssr.pillarbox.player.qos.models
 /**
  * Represents a generic event, which contains metrics about the current media stream.
  *
- * @property bandwidth The device-measured network bandwidth, in bytes per second.
- * @property bitrate The bitrate of the current stream, in bytes per second.
+ * @property bandwidth The device-measured network bandwidth, in bits per second.
+ * @property bitrate The bitrate of the current stream, in bits per second.
  * @property bufferDuration The forward duration of the buffer, in milliseconds.
  * @property playbackDuration The duration of the playback, in milliseconds.
  * @property playerPosition The position of the player, in milliseconds.
@@ -17,7 +17,7 @@ package ch.srgssr.pillarbox.player.qos.models
  */
 data class QoSEvent(
     val bandwidth: Long,
-    val bitrate: Int,
+    val bitrate: Long,
     val bufferDuration: Long,
     val playbackDuration: Long,
     val playerPosition: Long,
